+++ libmve/mvelibl.h
@@ -20,6 +20,8 @@
 #ifndef MVELIB_H_INCLUDED
 #define MVELIB_H_INCLUDED
 
+#include <cstdint>
+
 #include "SystemInterfaces.h"
 #if defined(__LINUX__)
 #include "lnxdsound.h"
@@ -50,8 +52,8 @@ void MVE_sndInit(ISoundDevice *lpDS);
 // Pan ranges from -10,000 (left full volume, right -100db), thru 0 (both full),
 // thru 10,000 (left -100db, right full volume).
 // The default value for volume and pan is zero.
-void MVE_dsbSetVolume(long lVolume);
-void MVE_dsbSetPan(long lPan);
+void MVE_dsbSetVolume(int32_t lVolume);
+void MVE_dsbSetPan(int32_t lPan);
 
 // Only call this function to configure software to work with a Super VGA
 //  mode if you do not have VESA support.
